+++ packages/core/utils/query.ts
@@ -1,14 +1,24 @@
 import { DatabaseInterface, Settings } from '../interface';
 import { format } from '@sqltools/plugins/formatter/utils';
+import multipleQueiesParse from './query/parse';
 
-export function parse(query = '') {
-  // @todo: use https://github.com/TeamSQL/SQL-Statement-Parser for better parsing
-  return query.split(/\s*;\s*(?=([^']*'[^']*')*[^']*$)/g).filter((v) => !!v && !!`${v}`.trim());
+export function parse(query: string, dialect: 'pg' | 'mysql' | 'mssql' = 'mysql', delimiter: string = ';'): string[] {
+  try {
+    return multipleQueiesParse(query.replace(/^[ \t]*GO;?[ \t]*$/gmi, ''), dialect, delimiter)
+  } catch (error) {
+    return query.split(/\s*;\s*(?=([^']*'[^']*')*[^']*$)/g).filter((v) => !!v && !!`${v}`.trim());
+  }
 }
 
 // @todo add some tests for this new function
 export function cleanUp(query = '') {
-  return query.replace(/('(''|[^'])*')|[\t\r\n]|(--[^\r\n]*)|(\/\*[\w\W]*?(?=\*\/)\*\/)/gmi, '')
+  return query.replace('\t', '  ')
+    .replace(/('(''|[^'])*')|(--[^\r\n]*)|(\/\*[\w\W]*?(?=\*\/)\*\/)/gmi, '')
+    .split(/\r\n|\n/gi)
+    .map(v => v.trim())
+    .filter(Boolean)
+    .join(' ')
+    .trim();
 }
 
 export function generateInsert(